Let consumers retune the ZMES websocket alert off its flapping max shape; bump to 0.13.11
The opt-in "ZMES Websocket Response Time" rule reduces with `max` over a
5-minute window against a 0.5s threshold. That shape cannot express
"sustained slowness" -- a single slow sample sets the reduced value for
the whole window, so the alert fires, holds for exactly 5 minutes as the
spike ages out of the window, and clears.

On bigserver that is not a hypothetical: over 90 days the rule produced
111 firings, and *every* excursion above 0.5s was exactly one minute
long. There has never been a second consecutive minute above threshold.
Baseline is 0.008s (p99 0.014s) with isolated 0.8-3.2s samples, so the
rule has fired 111 times and has never once indicated a real problem.

Widening the window would make this worse, not better, under a max
reducer -- it just holds the spike longer. The reducer itself has to
change. The new zmes_websocket_aggregation='mean' switches the query to
avg_over_time and the reduce stage to mean, which averages a spike away
across the window while still catching genuinely sustained latency;
zmes_websocket_threshold / _from / _for tune the rest.

All four parameters default to exactly the previous behavior, including
the word "maximum" in the annotation, so existing consumers see no
Terraform diff at all.

Also widen NutUps.runtime_minutes from int to float: the same class of
retuning on a UPS runtime alert needs fractional minutes, and the value
was only ever passed through as an alert threshold.

+++ grafana_cdktf_helpers/nut.py
@@ -25,7 +25,7 @@ class NutUps:
 
     def __init__(
         self, stack: 'BaseStack', upsname: str, folder_id: str, folder_uid: str,
-        runtime_minutes: int, has_output_voltage: bool = True,
+        runtime_minutes: float, has_output_voltage: bool = True,
         battery_voltage: int = 24, load_threshold: int = 75,
         add_rules: bool = True,
         status_no_data_state: str = 'NoData',
